+++ app/components/claim-form.tsx
@@ -2,7 +2,7 @@
 
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
-import { startTransaction } from "@/app/actions";
+import { startTransaction } from "@/api/actions";
 import { useFormState, useFormStatus } from "react-dom";
 import { useEffect, useState } from "react";
 import { Spinner } from "./ui/spinner";
