+++ next.config.mjs
@@ -4,8 +4,7 @@ const nextConfig = {
     remotePatterns: [
       {
         protocol: "https",
-        hostname: "i.ibb.co",
-        port: "",
+        hostname: "**",
       },
     ],
   },
